% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adapt.R
\name{adapt.dirichlet}
\alias{adapt.dirichlet}
\title{Adapt parameters for dirichlet node}
\usage{
\method{adapt}{dirichlet}(object, samples, total)
}
\arguments{
\item{samples}{the set of samples.}

\item{total}{the sum of the likelihood of the samples.}

\item{dirichlet.object}{the node object.}
}
\value{
the updated node
}
\description{
AdaptS a dirichlet node's parameters given a set of samples.
}
\keyword{internal}

