% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adapt.R
\name{adapt.flagprod}
\alias{adapt.flagprod}
\title{Adapt parameters for flagprod node}
\usage{
\method{adapt}{flagprod}(object, samples, total)
}
\arguments{
\item{object}{the node object.}

\item{samples}{the set of samples.}

\item{total}{the sum of the likelihood of the samples.}
}
\value{
the updated node
}
\description{
AdaptS a flagprod node's parameters given a set of samples.
}
\keyword{internal}

