% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adapt.R
\name{adapt.noisyor}
\alias{adapt.noisyor}
\title{Adapt parameters of noisy or node}
\usage{
\method{adapt}{noisyor}(object, samples, total)
}
\arguments{
\item{samples}{list of current samples}

\item{total}{numeric giving sum of likelihood of all samples}

\item{noisyor.object}{the node}

\item{debug}{boolean specifying if debug messages should be shown}
}
\value{
updated node
}
\description{
Adapt parameters of noisy or node given current samples
}
\keyword{internal}

