% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adapt.R
\name{adapt}
\alias{adapt}
\title{Adapt parameters}
\usage{
adapt(object, samples, total)
}
\arguments{
\item{object}{the node object.}

\item{samples}{the set of samples.}

\item{total}{the sum of the likelihood of the samples.}
}
\value{
the updated node
}
\description{
S3 generic function for adapting a node's parameters given a set of samples.
}
\keyword{internal}

