% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddEdge.R
\name{addEdge}
\alias{addEdge}
\title{Add an edge into/a parent to a node}
\usage{
addEdge(net, child, parent, params, params2 = NULL, cn = NULL,
  deepUpdate = F, dropImages = F)
}
\arguments{
\item{net}{The network.}

\item{child}{The index of the node to add
a parent to.}

\item{parent}{The index of the node to add as a parent.}

\item{params}{The new parameters for the child.}

\item{params2}{The disambiguation parameters to use for the child. This
is used only if the child is a multinomial noisy-or node. If the child is
multinomial noisy-or and this is NULL a weighted mutex disambiguation
matrix is created.}

\item{cn}{A RODBC connection. Required if performing deep update.}

\item{deepUpdate}{Specifies whether a deep update should be performed.}

\item{dropImages}{Specifies if images should be dropped on a deepUpdate.}

\item{params}{The parameters to add to the child for the parent.}
}
\value{
Updated network.
}
\description{
Add an edge into/a parent to a node. Node must be Noisy-Or or Dirichlet.
}

