% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddEdge.R
\name{addEdgeDirichlet}
\alias{addEdgeDirichlet}
\title{Add an edge into/a parent to a dirichlet node}
\usage{
addEdgeDirichlet(net, child, parent, params)
}
\arguments{
\item{net}{The network.}

\item{child}{The index of the dirichlet node to add
a parent to.}

\item{parent}{The index of the node to add as a parent.}

\item{params}{The new parameters for the child.}
}
\value{
Updated network.
}
\description{
Add an edge into/a parent to a dirichlet node
}
\keyword{internal}

