% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddEdge.R
\name{addEdgeNoisyOr}
\alias{addEdgeNoisyOr}
\title{Add an edge into/a parent to a noisy-or node}
\usage{
addEdgeNoisyOr(net, child, parent, params, params2 = NULL)
}
\arguments{
\item{net}{The network.}

\item{child}{The index of the noisy-or node to add a parent to.}

\item{parent}{The index of the node to add as a parent.}

\item{params}{The parameters to add to the child for the parent. This
should be a matrix with as many rows as the parent node has
non-zero values (I.e. one less than the number of values of the
parent).}

\item{params2}{The disambiguation parameters to use for the child. This
is used only if the child is multinomial. If the child is multinomial and this is NULL
a weighted mutex disambiguation matrix is created.}
}
\description{
Add an edge into/a parent to a noisy-or node
}
\keyword{internal}

