% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{addInformationToMetaData}
\alias{addInformationToMetaData}
\title{Add information to metadata table}
\usage{
addInformationToMetaData(cn, metatable, tableinfo, net, table_name,
  table_number, number_of_nodes, max_length_of_table)
}
\arguments{
\item{cn}{A RODBC connection}

\item{metatable}{The metatable to add information to}

\item{tableinfo}{Table information object for network, must have valid data family name.}

\item{net}{The Bayesnet object}

\item{table_name}{The name of the table information is being added about}

\item{table_number}{Number of the result table}

\item{number_of_nodes}{The number of nodes in the result table}

\item{max_length_of_table}{The maximum number of nodes in the result table}
}
\description{
Add information to metadata table
}
\keyword{internal}

