% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{addInformationToNetworks}
\alias{addInformationToNetworks}
\title{Add network information to networks table}
\usage{
addInformationToNetworks(cn, tableinfo, layout)
}
\arguments{
\item{cn}{A RODBC Connection}

\item{tableinfo}{A table information object for the network}

\item{layout}{The layout matrix for the network}
}
\description{
Add network information to networks table
}
\keyword{internal}

