% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddNode.R
\name{addNode}
\alias{addNode}
\title{Add node to network}
\usage{
addNode(net, type, values, name, parents, params, params2 = NULL,
  miscString = NULL, cn = NULL, deepUpdate = F, dropImages = F)
}
\arguments{
\item{net}{the current network}

\item{type}{string specifying type of node to add. Valid values are:
'noisyor', 'dirichlet', 'logand', 'logor', logprod', 'flag',
and 'flagprod'.}

\item{values}{vector of strings specifying the names of the values of the
node. Note that all nodes must have at least two values, and noisy-or and
logical-and type nodes can only have two values.}

\item{name}{string specifying the name of the node.}

\item{parents}{vector of integers specifying the indices of the parents of
node. Note that logical-and nodes must have exactly two parents.}

\item{params}{matrix specifying parameters of node.}

\item{params2}{matrix specifying secondary parameters of node.}

\item{miscString}{a string specifying additional settings.
Currently used only for multinomial noisy-or nodes, in
which case it should be 'T' if the secondary parameters
matrix should be adaptable, and 'F' otherwise.}

\item{cn}{A RODBC connection. Required if performing deep update.}

\item{deepUpdate}{Specifies whether a deep update should be performed.}

\item{dropImages}{Specifies if images should be dropped on a deepUpdate.}
}
\value{
new Bayesian network with node added.
}
\description{
Adds a new node of to the network.
NOTE: RESULTS WILL BE CLEARED. If you want to have access to past results you must copy and
store them elsewhere before calling this function.
NOTE: ADDING A NODE CAN RESULT IN A NEW DATA TABLE BEING CREATED.
}

