% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddNode.R
\name{addOrRestoreVariable}
\alias{addOrRestoreVariable}
\title{Add or restore variable to data tables}
\usage{
addOrRestoreVariable(cn, tableinfo, newNode, newvalues)
}
\arguments{
\item{cn}{An RODBC connection}

\item{tableinfo}{A table information object}

\item{newNode}{The name of the new variable}

\item{newValues}{The number of values the new variable has}
}
\value{
The update tableinfo object
}
\description{
Add or restore variable to data tables.
Checks if the new variable to be added is already present in the master data tables.
If so, and it is valid (the desired values match stored values) it is 'restored' which
simply means adding it to the datamap.
Otherwise the variable is added to the master data tables and then to the map.
}
\keyword{internal}

