% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddNode.R
\name{addVariableToDataMaster}
\alias{addVariableToDataMaster}
\title{Add variable to master data tables}
\usage{
addVariableToDataMaster(cn, tableinfo, name, values)
}
\arguments{
\item{cn}{connection}

\item{tableinfo}{Table information object}

\item{name}{Name of new node}

\item{values}{Number of values of new node}
}
\description{
Add variable to master data tables
}
\keyword{internal}

