% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Analysis.R
\name{analyzer}
\alias{analyzer}
\title{Create an analysis object}
\usage{
analyzer(rootCauses, defects)
}
\arguments{
\item{rootCauses}{a matrix of two columns specifying the
variables/value combinations that represent root causes.}

\item{defects}{a matrix of two columns specifying the
variables/value combinations that represent defects.}
}
\value{
An analysis object
}
\description{
Create an analysis object. If we have three variables:
 Rain     : {None,Light,Heavy}
 Sprinker : {Off,On}
 WetGrass : {False,True}
And we consider wet grass to be a defect, and rain or the
sprinkler to be potential root causes, then our matrices
would be:
 Root Cause Matrix:
   1 , 2
   1 , 3
   2 , 2
 Defect Matrix:
   3 , 2
}

