% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bayesnet.R
\name{bayesnet}
\alias{bayesnet}
\title{Bayes net constructor}
\usage{
bayesnet(name, dbinfo = dbmsinfo(), connection_string = NULL,
  external_path = NULL, table_length = 200)
}
\arguments{
\item{name}{The name of the new network}

\item{dbinfo}{A database information object, returned from intializeDatabase. If not provided,
a default object will be created, with networks="networks", nodes="nodes", log="log", and metadata="metadata"}

\item{connection_string}{Connection string to database.}

\item{external_path}{Path to external inference executables}

\item{table_length}{The number of variables to have in data and result tables.}
}
\value{
new Bayesian network
}
\description{
Bayes net constructor. Nodes are added using 'addNode' function.
The net object includes the layout field, which you should edit to control
plot layouts of the network. It is at $layout and is
a matrix specifying the plot coordinates of the nodes.
If you wish to add external_path and connection_string at a later point, you can
access these at $inf_manager$external_path and $inf_manager$connection_string. If they
are null, you will not be able to run the external inference engine and other
external apps.
}

