% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/noisyor.R
\name{calculate.pindices.noisyor}
\alias{calculate.pindices.noisyor}
\title{Calculate parameter jump table for noisy or node}
\usage{
calculate.pindices.noisyor(pvals)
}
\arguments{
\item{pvals}{a vector specifying the number of values each parent has.}
}
\value{
a vector specifying the first row in the node's parameters
corresponding to each parent
}
\description{
Calculate jump table specifying the first row in the node's parameters
corresponding to each parent for a noisy or node
}
\keyword{internal}

