% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_posteriori.R
\name{calculate.posteriori}
\alias{calculate.posteriori}
\title{Calcalate posteriori distribution for node}
\usage{
calculate.posteriori(node, samples, evidence)
}
\arguments{
\item{node}{the node}

\item{samples}{list of samples and their likelihood given evidence}

\item{evidence}{vector of integers specifying known values of nodes. This
can include NA values for unknown variables.}
}
\value{
posteriori distribution for node
}
\description{
Calcalate posteriori distribution for node given samples and evidence
}
\keyword{internal}

