% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prob_functions.R
\name{calculate.row}
\alias{calculate.row}
\title{Calculate the row in parameters for given parent values}
\usage{
calculate.row(cur, max)
}
\arguments{
\item{cur}{vector specifying values of parents}

\item{max}{vector specifying number of values of parents}
}
\value{
integer specifying row index
}
\description{
Calculate the row in parameters for given parent values
}
\keyword{internal}

