% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadResults.R
\name{checkResultsForItem}
\alias{checkResultsForItem}
\title{Check results for item.}
\usage{
checkResultsForItem(connection, net, item)
}
\arguments{
\item{connection}{An open RODBC connection}

\item{net}{The network, with a valid results table name in
the resultsTable field - so network tables must have been
initialized.}

\item{item}{The item to retrieve.}
}
\value{
Boolean specifying if item is present
}
\description{
Check results for item.
}

