% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{clearAllData}
\alias{clearAllData}
\title{Remove all entries from data tables for a network.}
\usage{
clearAllData(cn, net, ignore_warnings = F)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{net}{The network.}
}
\description{
Remove all entries from data tables for a network. If multiple networks are using the data tables the
function will fail with a message unless ignore_warnings is set to TRUE.
}

