% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{clearAllImages}
\alias{clearAllImages}
\title{Remove all entries from image table for a network.}
\usage{
clearAllImages(cn, net)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{net}{The network.}
}
\description{
Remove all entries from image table for a network.
}

