% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{clearAllResults}
\alias{clearAllResults}
\title{Remove all entries from result tables for a network.}
\usage{
clearAllResults(cn, net)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{net}{The network.}
}
\description{
Remove all entries from result tables for a network.
}

