% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LocationTableManagement.R
\name{clearDatabase}
\alias{clearDatabase}
\title{Clear all ABN tables in a database.}
\usage{
clearDatabase(cn, dbinfo)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{dbInfo}{A database information object returned from intializeDatabase}
}
\description{
Clear all ABN tables in a database.
}

