% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/InferenceManagement.R
\name{clone_inference_management}
\alias{clone_inference_management}
\title{Clone inference management object}
\usage{
clone_inference_management(new_name, other)
}
\arguments{
\item{new_name}{The name of the new, cloned network}

\item{other}{The inference management object to be cloned}
}
\value{
A new inference management object
}
\description{
Clone inference management object
}
\keyword{internal}

