% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bayesnet.R
\name{cloneBayesnet}
\alias{cloneBayesnet}
\title{Bayes net cloner}
\usage{
cloneBayesnet(net, name)
}
\arguments{
\item{net}{Network to clone}

\item{name}{Name for new network}
}
\value{
new Bayesian network
}
\description{
Copies passed net'Bayesnet's nodes and layout, and gives
network the passed name.
Also copies the data master tables names and the variable map onto these tables.
}

