% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TableInfo.R
\name{cloneTableInfo}
\alias{cloneTableInfo}
\title{Clone table information object}
\usage{
cloneTableInfo(newname, tableinfo)
}
\arguments{
\item{newname}{The name of the Bayesian network the new table information object will
be associated with.}

\item{tableinfo}{The table information object to be cloned.}
}
\value{
A new table information object
}
\description{
Clone table information object
}
\keyword{internal}

