% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RODBCWrapper.R
\name{closeODBCConnection}
\alias{closeODBCConnection}
\title{Close an open ODBC connection}
\usage{
closeODBCConnection(cn)
}
\arguments{
\item{cn}{An open RODBC connection}
}
\description{
Close an open ODBC connection
}

