% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CompleteTest.R
\name{CompletePackageTest}
\alias{CompletePackageTest}
\title{Complete Package Test}
\usage{
CompletePackageTest(clear = T,
  connection_string = getMySQLConnectionString(),
  external_path = "C:\\\\Users\\\\User\\\\Documents\\\\Visual Studio 2013\\\\Projects\\\\abnInference\\\\Release",
  seed = NULL)
}
\description{
Test all functionality in package.
Work on this function is ongoing.
}

