% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{createDataTable}
\alias{createDataTable}
\title{Create Data table}
\usage{
createDataTable(cn, name_base, table_number, number_of_nodes,
  max_length_of_table)
}
\arguments{
\item{cn}{A RODBC Connection}

\item{name_base}{Base name for the data tables}

\item{table_number}{Number of data table}

\item{number_of_nodes}{The number of nodes in the data table}

\item{max_length_of_table}{The maximum number of nodes in the table}
}
\value{
The data table name
}
\description{
Create Data table (singular).
}
\keyword{internal}

