% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{createDataTables}
\alias{createDataTables}
\title{Create Data tables}
\usage{
createDataTables(cn, tableinfo, net)
}
\arguments{
\item{cn}{A RODBC Connection}

\item{tableinfo}{Table information object for network}

\item{net}{The Bayesnet object}
}
\value{
Vector of data table names
}
\description{
Create Data tables
}
\keyword{internal}

