% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{createDataView}
\alias{createDataView}
\title{Create Data view}
\usage{
createDataView(cn, datatables, datamap, name_base, table_number,
  number_of_nodes, max_length_of_table)
}
\arguments{
\item{cn}{A RODBC Connection}

\item{datatables}{The names of the data tables associated with the network}

\item{datamap}{Network's datamap}

\item{name_base}{Base name for the data views}

\item{table_number}{Number of the data view}

\item{number_of_nodes}{The number of nodes in the data view}

\item{max_length_of_table}{The maximum number of nodes in the view}
}
\value{
The data view name
}
\description{
Create Data view (singular).
}
\keyword{internal}

