% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{createDataViews}
\alias{createDataViews}
\title{Create Data views}
\usage{
createDataViews(cn, tableinfo)
}
\arguments{
\item{cn}{A RODBC Connection}

\item{tableinfo}{Table information object for network}
}
\value{
Vector of data view names
}
\description{
Create Data views
}
\keyword{internal}

