% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{createImageTable}
\alias{createImageTable}
\title{Create image table}
\usage{
createImageTable(cn, tableinfo)
}
\arguments{
\item{cn}{A RODBC Connection}

\item{tableinfo}{Table information object for network}
}
\value{
Image table name
}
\description{
Create image table
}
\keyword{internal}

