% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LocationTableManagement.R
\name{createLogTable}
\alias{createLogTable}
\title{Create a database's log table}
\usage{
createLogTable(cn, table)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{table}{The string to use as the name of the log table.
This should be unique in the database.}
}
\description{
Create a database's log table
}
\keyword{internal}

