% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LocationTableManagement.R
\name{createMetaresultsTable}
\alias{createMetaresultsTable}
\title{Create a database's metaresults table}
\usage{
createMetaresultsTable(cn, table)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{table}{The string to use as the name of the result data
table. This should be unique in the database.}
}
\description{
Create a database's metaresults table
}
\keyword{internal}

