% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LocationTableManagement.R
\name{createNetworksTable}
\alias{createNetworksTable}
\title{Create a database's network table}
\usage{
createNetworksTable(cn, table)
}
\arguments{
\item{cn}{An open RODBC connection}

\item{table}{The string to use as the name of the
network table. This should be unique in the database.}
}
\description{
Create a database's network table
}
\keyword{internal}

