% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LocationTableManagement.R
\name{createNodesTable}
\alias{createNodesTable}
\title{Create a database's nodes table}
\usage{
createNodesTable(cn, table, networks)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{table}{The string to use as the name of the nodes
table. This should be unique in the database.}

\item{networks}{The name of the network table in
the database.}
}
\description{
Create a database's nodes table
}
\keyword{internal}

