% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{createResultTable}
\alias{createResultTable}
\title{Create Result table}
\usage{
createResultTable(cn, name_base, table_number, net, max_length_of_table)
}
\arguments{
\item{cn}{A RODBC Connection}

\item{name_base}{Base name for the result table}

\item{table_number}{Number of the result table}

\item{net}{The Bayesnet object}

\item{max_length_of_table}{The maximum number of nodes in the result table}
}
\value{
The result table name
}
\description{
Create Result table (singular).
}
\keyword{internal}

