% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{createResultTables}
\alias{createResultTables}
\title{Create Result tables}
\usage{
createResultTables(cn, net, tableinfo)
}
\arguments{
\item{cn}{A RODBC Connection}

\item{net}{The Bayesnet object (perhaps with out of date tableinfo)}

\item{tableinfo}{Table information object for network}
}
\value{
Vector of result table names
}
\description{
Create Result tables
}
\keyword{internal}

