% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TableInfo.R
\name{createTableInfo}
\alias{createTableInfo}
\title{Create table information object}
\usage{
createTableInfo(netname, table_length = 200, dbinfo)
}
\arguments{
\item{netname}{The name of the Bayesian network associated with this object.}

\item{table_length}{The number of variables to have in data and result tables.}

\item{dbinfo}{A database information object.}
}
\value{
A table information object
}
\description{
Create table information object. This holds information on all database tables
associated with the network.
}

