% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbmsinfo.R
\name{dbmsinfo}
\alias{dbmsinfo}
\title{Create database information object.}
\usage{
dbmsinfo(networks = "networks", metadata = "metadata", nodes = "nodes",
  log = "log", metaresults = "metaresults")
}
\arguments{
\item{networks}{The name of the networks table.}

\item{metadata}{The name of the metadata table.}

\item{nodes}{The name of the nodes table.}

\item{log}{The name of the log table.}

\item{metaresults}{The name of the metaresults table.}
}
\value{
A database information object.
}
\description{
Create database information object. This contains names for all location tables.
}

