% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DebugWrappers.R
\name{debugClearDatabase}
\alias{debugClearDatabase}
\title{A debug wrapper function for clearing the database.}
\usage{
debugClearDatabase()
}
\description{
A debug wrapper function for clearing the database. Should be placed in a separate package.
}
\keyword{internal}

