% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DebugWrappers.R
\name{debugInitializeDatabase}
\alias{debugInitializeDatabase}
\title{A debug wrapper function for initilizing the database.}
\usage{
debugInitializeDatabase()
}
\description{
A debug wrapper function for initilizing the database. Should be placed in a separate package.
}
\keyword{internal}

