% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddNode.R
\name{deepUpdateOnNodeAdded}
\alias{deepUpdateOnNodeAdded}
\title{Perform a deep update after adding a node}
\usage{
deepUpdateOnNodeAdded(cn, net, dropImages = F)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{net}{The network.}

\item{dropImages}{Whether images should be dropped or updated.}
}
\value{
The updated network
}
\description{
Update the images and tables of a network after adding a node.
}
\keyword{internal}

