% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RemoveNode.R
\name{deepUpdateOnNodeRemoved}
\alias{deepUpdateOnNodeRemoved}
\title{Perform a deep update after removing a node.}
\usage{
deepUpdateOnNodeRemoved(cn, net, oldIndex, oldValues, dropImages = F)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{net}{The network.}

\item{oldIndex}{The index of the node that has been removed.}

\item{oldValues}{The number of values the dropped node had.}

\item{dropImages}{Whether images should be dropped or updated.}

\item{zeroIndexed}{Whether the oldIndex is zero or one indexed.}
}
\value{
The updated network
}
\description{
Update a network in the database after a nodeis removed.
Note that updating images is permitted only when the removed node has no children
in the network or in any network images.
Use deepUnpdateOnEdgeRemoved to remove egdes in images.
NB This function requires that there not exist multiple multiple tables with the same name as the master data
table in different data bases / data schema of the data management system.
}
\keyword{internal}

