% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{dirichlet}
\alias{dirichlet}
\title{Dirichlet node constructor}
\usage{
dirichlet(index, values, name = as.character(index), parents = NULL,
  pvals = NULL, params = NULL, params2 = NULL)
}
\arguments{
\item{index}{index of node}

\item{values}{vector of string giving names of values of node}

\item{name}{string giving name of node}

\item{parents}{vector specifying indices of parents of node}

\item{pvals}{vector specifying number of values for each parent of node}

\item{params}{matrix specifying parameters of node}
}
\value{
constructed dirichlet node
}
\description{
Dirichlet node constructor
}
\keyword{internal}

