% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DropResults.R
\name{dropResults}
\alias{dropResults}
\title{Delete results}
\usage{
dropResults(cn, net, first)
}
\arguments{
\item{cn}{An open RODBC connection}

\item{net}{The network}

\item{first}{The first item to delete. All subsequent items will be deleted.}
}
\description{
Delete results from a network's results table.
}
\keyword{internal}

