% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddEdge.R
\name{find.loop}
\alias{find.loop}
\title{Find a loop in graph}
\usage{
find.loop(edges, from, target)
}
\arguments{
\item{edges}{the edge matrix}

\item{from}{the nodes we are looking for routes to the target from}

\item{target}{the node which a route to proves there is a loop}
}
\value{
Boolean specifying whether loop exists or not
}
\description{
Find a loop in graph
}
\keyword{internal}

