% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flag.R
\name{flag}
\alias{flag}
\title{Constructor for flag node}
\usage{
flag(index, name, values, parents, pvals, params, params2)
}
\arguments{
\item{index}{integer giving the index of the node}

\item{name}{string giving name of node}

\item{values}{vector of strings giving names of values of node.
This must be length 2.}

\item{pvals}{vector specifying number of values for each parent of node.
This must be length 1}

\item{params}{matrix specifying parameters of node: The relevant value of the
parent as a 1x1 matrix.}
}
\value{
constructed flag node
}
\description{
Constructor for flag node
}
\keyword{internal}

