% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.dirichlet}
\alias{gen.dirichlet}
\title{Generate random value for a dirichlet node}
\usage{
\method{gen}{dirichlet}(object, pvals, aux, met)
}
\arguments{
\item{pvals}{vector specifying the values taken by the node's parents.}

\item{dirichlet.object}{the node.}
}
\value{
the generated value.
}
\description{
Generate a random value for a dirichlet node given the values that the
node's parents take.
}
\keyword{internal}

