% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.flagprod}
\alias{gen.flagprod}
\title{Generate random value for a flagprod node}
\usage{
\method{gen}{flagprod}(object, pvals, aux, met)
}
\arguments{
\item{object}{the node.}

\item{pvals}{vector specifying the values taken by the
node's parents.}
}
\value{
the generated value.
}
\description{
Generate a random value for a flagprod node given the
values that the node's parents take.
}
\keyword{internal}

