% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.noisyor}
\alias{gen.noisyor}
\title{Generate random value for a noisy or node}
\usage{
\method{gen}{noisyor}(object, pvals, aux, met)
}
\arguments{
\item{pvals}{the values taken by the node's parents.}

\item{noisyor}{the node.}
}
\value{
the generated value.
}
\description{
Generate a random value for a noisy or node given the values
that the node's parents take.
}
\keyword{internal}

