% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gen.R
\name{gen}
\alias{gen}
\title{Generate random value}
\usage{
gen(object, pvals, aux, met = "aux")
}
\arguments{
\item{object}{the node object.}

\item{pvals}{the values taken by the node's parents.}

\item{met}{"nat" or "aux": Generate from true or effective parents
of noisy or nodes}
}
\value{
the generated value.
}
\description{
S3 generic function for generating a random value given
a nodes parents take specific values.
}
\keyword{internal}

