% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prob_functions.R
\name{genFromRow}
\alias{genFromRow}
\title{Generate value from parameters row}
\usage{
genFromRow(params)
}
\arguments{
\item{params}{row vector from parameters matrix to be used}
}
\value{
integer specifying generated value
}
\description{
Generate value from parameters row
}
\keyword{internal}

