% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddNode.R
\name{get.values.bayesnet}
\alias{get.values.bayesnet}
\title{Get the number of values for the specified nodes}
\usage{
get.values.bayesnet(net, pars)
}
\arguments{
\item{net}{the network}

\item{pars}{vector of indices of the nodes to get information about}
}
\value{
vector of integers specifying the number of values for the
specified nodes
}
\description{
Get the number of values for the specified nodes
}
\keyword{internal}

