% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getmutualinformation.R
\name{getChildren}
\alias{getChildren}
\title{Get the children of a node}
\usage{
getChildren(node, net)
}
\arguments{
\item{node}{The node whose children to fetch}

\item{net}{The network}
}
\value{
A vector of indices of the children of the node
}
\description{
Get the children of a node
}
\keyword{internal}

