% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RewindReadapt.R
\name{getClosestToTime}
\alias{getClosestToTime}
\title{Get a value from a time}
\usage{
getClosestToTime(cn, table, time, col, timeCol = "time", prior = T,
  secondaryOrder = NULL)
}
\arguments{
\item{cn}{An open RODBC connection}

\item{table}{The table name}

\item{time}{The time the entry should be closest but prior to/after}

\item{col}{The column of the entry to return}

\item{timeCol}{The column where the time stamp to compare is.}

\item{prior}{If the compared value must be prior to (T) or after (F)}

\item{secondaryOrder}{A secondary column to order on.}
}
\description{
Get the value in column col that is in the row whose
timeCol timestamp is closest but prior to time from table.
}
\keyword{internal}

