% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DebugWrappers.R
\name{getDebugODBCConnection}
\alias{getDebugODBCConnection}
\title{A debug wrapper function for connecting to the database.}
\usage{
getDebugODBCConnection()
}
\description{
A debug wrapper function for connecting to the database. Should be placed in a separate package.
}
\keyword{internal}

