% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getmutualinformation.R
\name{getDescendents}
\alias{getDescendents}
\title{Get Descendents}
\usage{
getDescendents(node, net)
}
\arguments{
\item{node}{Index of node}

\item{net}{The Bayesnet object}
}
\value{
The indices of descendents to the node in the net
}
\description{
Get Descendents
}
\keyword{internal}

