% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_edges.R
\name{getEdges}
\alias{getEdges}
\title{Get the edges present in a network}
\usage{
getEdges(net)
}
\arguments{
\item{net}{the network}
}
\value{
a vector specifying the edges present, with odd indexed values giving
the parent indices, and even number indices giving the child index.
}
\description{
Get the edges present in a network
}

