% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LocationActions.R
\name{getLogEntries}
\alias{getLogEntries}
\title{Get log entries}
\usage{
getLogEntries(cn, log = "log")
}
\arguments{
\item{cn}{An open RODBC connection}

\item{log}{The log table in the database}
}
\value{
A data frame of the log entries
}
\description{
Get log entries
}

