% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getmutualinformation.R
\name{getMutualInformation}
\alias{getMutualInformation}
\title{Estimate mutual information}
\usage{
getMutualInformation(net, evidence, nodes, envNodes, weights = rep(1,
  length(nodes)), num = 1000, algorithm = "IS", algoSpecific = "nat",
  debug = FALSE)
}
\arguments{
\item{net}{the network}

\item{evidence}{vector of integers specifying known values of nodes. This
can include NA values for unknown variables.}

\item{nodes}{The nodes to estimate informative
environment variables for.}

\item{envNodes}{The environment variables.}

\item{weights}{A weight vector to combine
results for weighted overall information.}

\item{num}{integer specifying the number of samples to used in inference}

\item{algorithm}{A string specifying a valid sampling algorithm.}

\item{algoSpecific}{Required settings for sampling algorithm.}

\item{debug}{For internal use only.}
}
\value{
matrix with each row giving
mutual information of node with environment
variables. An additional row exists for the
weighted combined information if weights is
not NULL.
}
\description{
Estimate mutual information on network
given evidence. Note the mutual information of a variable with itself
is the entropy of that variable (not necessarily 1).
}

