% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DebugWrappers.R
\name{getMySQLConnectionString}
\alias{getMySQLConnectionString}
\title{A debug wrapper function for connecting to the mySQL database.}
\usage{
getMySQLConnectionString()
}
\description{
A debug wrapper function for connecting to the mySQL database. Should be placed in a separate package.
}
\keyword{internal}

