% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkSerialization.R
\name{getNetwork}
\alias{getNetwork}
\title{Retrieve a network from a connection}
\usage{
getNetwork(cn, name, connection_string = NULL, external_path = NULL,
  dbinfo = dbmsinfo())
}
\arguments{
\item{cn}{An open RODBC connection}

\item{name}{The name of the network to retrieve}

\item{connection_string}{Connection string to database.}

\item{external_path}{Path to external inference executables}

\item{dbinfo}{A database information object}
}
\value{
The retrieved bayesnet
}
\description{
Retrieve a network from a connection
}

