% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RODBCWrapper.R
\name{getODBCConnection}
\alias{getODBCConnection}
\title{Open an ODBC connection}
\usage{
getODBCConnection(connection_string)
}
\arguments{
\item{connection_string}{A valid ODBC connection string}
}
\value{
An RODBC connection object
}
\description{
Open an ODBC connection
}

