% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getParameters.R
\name{getParameters1}
\alias{getParameters1}
\title{S3 generic function for getting the parameters from a node.}
\usage{
getParameters1(object)
}
\arguments{
\item{object}{the node object.}
}
\value{
a vector giving the parameters
}
\description{
S3 generic function for getting the parameters from a node.
}
\keyword{internal}

