% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getRequiredAuxiliaryPlaces.R
\name{getRequiredAuxiliaryPlaces}
\alias{getRequiredAuxiliaryPlaces}
\title{S3 generic function for serialization}
\usage{
getRequiredAuxiliaryPlaces(node)
}
\arguments{
\item{node}{The node object.}
}
\value{
A vector giving the parameters
}
\description{
S3 generic function for getting the number of additional
sample slots required by a node
parameters from a node.
}
\keyword{internal}

