% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DebugWrappers.R
\name{getSQLServerConnectionString}
\alias{getSQLServerConnectionString}
\title{A debug wrapper function for connecting to the SQL Server database.}
\usage{
getSQLServerConnectionString()
}
\description{
A debug wrapper function for connecting to the SQL Server database. Should be placed in a separate package.
}
\keyword{internal}

