% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddNode.R
\name{getVariableNamesAndValuesInMetaTables}
\alias{getVariableNamesAndValuesInMetaTables}
\title{Get names and number of values of variables in meta tables}
\usage{
getVariableNamesAndValuesInMetaTables(cn, tablelist, tablename, colname)
}
\arguments{
\item{cn}{A RODBC connection}

\item{tablelist}{List of table names (data or result) to look for}

\item{tablename}{Meta table to use}

\item{colname}{Column in meta table specifying tables of interest}
}
\value{
A list of two vectors, containing variable names and numbers of values
}
\description{
Get names and number of values of variables in meta tables
}
\keyword{internal}

