% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddNode.R
\name{getVariableNamesAndValuesInResultTables}
\alias{getVariableNamesAndValuesInResultTables}
\title{Get names and number of values of variables in result tables}
\usage{
getVariableNamesAndValuesInResultTables(cn, tableinfo)
}
\arguments{
\item{cn}{A RODBC connection}

\item{tableinfo}{A table information object}
}
\value{
A list of two vectors, containing variable names and numbers of values
}
\description{
Get names and number of values of variables in result tables
}
\keyword{internal}

