% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/InferenceManagement.R
\name{inference_management}
\alias{inference_management}
\title{Create inference management object}
\usage{
inference_management(name, external_path = NULL, connection_string = NULL)
}
\arguments{
\item{name}{Network name}

\item{external_path}{Path to external inference executables}

\item{connection_string}{Connection string to database.}
}
\value{
Inference management object
}
\description{
Create inference management object
}

