% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LocationTableManagement.R
\name{initializeDatabase}
\alias{initializeDatabase}
\title{Create required tables for a database.}
\usage{
initializeDatabase(cn, networks = "networks", nodes = "nodes",
  log = "log", metadata = "metadata", metaresults = "metaresults")
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{networks}{The desired name of the database's network table.}

\item{nodes}{The desired name of the database's nodes table.}

\item{log}{The desired name of the database's log table.}

\item{metadata}{The name of the database's metadata table.}

\item{metaresults}{The name of the database's metaresults table.}
}
\value{
A database information object
}
\description{
Create required tables for a database.
}

