% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddEdge.R
\name{isDAG}
\alias{isDAG}
\title{Check that graph is a dag}
\usage{
isDAG(net)
}
\arguments{
\item{net}{The Bayesian network object}
}
\value{
Boolean specifying if network a DAG
}
\description{
Check that graph is a dag. Note igraph::is.dag is buggy.
}
\keyword{internal}

