% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logand.R
\name{logand}
\alias{logand}
\title{Constructor for logical and node}
\usage{
logand(index, name, values, parents, pvals, params, params2)
}
\arguments{
\item{index}{integer giving the index of the node}

\item{name}{string giving name of node}

\item{values}{vector of strings giving names of values of node}

\item{parents}{vector specifying indices of parents of node}

\item{pvals}{vector specifying number of values for each parent of node}

\item{params}{matrix specifying values of parents for logical-and.}
}
\value{
constructed logand node
}
\description{
Constructor for logical and node
}
\keyword{internal}

