% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadData.R
\name{make_v_Strings}
\alias{make_v_Strings}
\title{Make V-Strings}
\usage{
make_v_Strings(number, base = 0)
}
\arguments{
\item{number}{The number of v-strings to make}

\item{base}{The base to add to the numbers}
}
\value{
A string of the form "v_1,v_2,...,v_number".
}
\description{
Make V-Strings
}
\keyword{internal}

