% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadResults.R
\name{make_vv_Strings}
\alias{make_vv_Strings}
\title{Make variable-value strings}
\usage{
make_vv_Strings(net)
}
\arguments{
\item{net}{The bayesnet object}
}
\value{
A string of the form "v_1_1,v_1_2,v_2_1,...,v_number_value".
}
\description{
Make variable-value strings
}
\keyword{internal}

