% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ManageExternalInference.R
\name{makeCommandArguments}
\alias{makeCommandArguments}
\title{Make command arguments}
\usage{
makeCommandArguments(settings)
}
\arguments{
\item{settings}{Settings list from makeExternalInferenceSettings}
}
\value{
command line string for starting external inference engine
}
\description{
Make command arguments.
}
\keyword{internal}

