% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DebugWrappers.R
\name{makeDebugSettings}
\alias{makeDebugSettings}
\title{A debug wrapper function for creating inference settings}
\usage{
makeDebugSettings(net, SystemID)
}
\description{
A debug wrapper function for creating inference settings. Should be placed in a separate package.
}
\keyword{internal}

