% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ManageExternalInference.R
\name{makeExternalInferenceSettings}
\alias{makeExternalInferenceSettings}
\title{Make external inference settings}
\usage{
makeExternalInferenceSettings(net, SystemID, Number, NextItem, Update,
  AssumeIntegrity, StoreImages, Compression_Setting, Compression_Value1,
  Compression_Value2, Pause, DBMS_Timeout)
}
\arguments{
\item{net}{A Bayesian network object}

\item{SystemID}{A unique ID for the external inference engine session.}

\item{Number}{The number of samples per inference case}

\item{NextItem}{The first item to start from. (0-Find Latest)}

\item{Update}{Whether network parameters should be updated.}

\item{AssumeIntegrity}{Specifies if engine can assume no other processes will change
database records while running.}

\item{StoreImages}{Whether images should be stored.}

\item{Compression_Setting}{'None', 'Until' or 'All'. None means compression value 2 is
used through out. All means compression value 1 is used through out. Until will use
value 1 until the process catches up with the number of entries in the data table, and
value 2 after that.}

\item{Compression_Value1}{The number of items to process during an initial catch up}

\item{Compression_Value2}{The number of items to process after any initial catch up}

\item{Pause}{The duration the external inference engine will wait before rechecking
if new data has appeared in the data table.}

\item{DBMS_Timeout}{Currently unused}
}
\value{
Inference settings object to be passed to makeCommandArguments
}
\description{
Make external inference settings
}
\keyword{internal}

