% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/noisyor.R
\name{makeMutex}
\alias{makeMutex}
\title{Make a mutual exclusion disambiguation matrix}
\usage{
makeMutex(parents, vals)
}
\arguments{
\item{parents}{A scalar specifying how many parents the node has.}

\item{vals}{A scalar specifying how many values the node has.}
}
\description{
Make a mutual exclusion disambiguation matrix
}

