% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/noisyor.R
\name{makeWeightedMutex}
\alias{makeWeightedMutex}
\title{Make a weighted mutual exclusion disambiguation matrix}
\usage{
makeWeightedMutex(parents, vals)
}
\arguments{
\item{parents}{A scalar specifying how many parents the node has.}

\item{vals}{A scalar specifying how many values the node has.}
}
\description{
Make a weighted mutual exclusion disambiguation matrix
}

