% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Safe_Min_Max.R
\name{my.max}
\alias{my.max}
\title{NULL-safe Maximizer}
\usage{
my.max(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
largest value in v, or 0 if v is NULL
}
\description{
Maximizer that return 0 if object is NULL
}
\keyword{internal}

