% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Safe_Min_Max.R
\name{my.min}
\alias{my.min}
\title{NULL-safe Minimizer}
\usage{
my.min(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
smallest value in v, or 0 if v is NULL
}
\description{
Minimizer that return 0 if object is NULL
}
\keyword{internal}

