% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{networksUsingDataTable}
\alias{networksUsingDataTable}
\title{Find networks associated with a master data table.}
\usage{
networksUsingDataTable(cn, tableinfo)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{dataTable}{The name of the master data table.}

\item{networks}{The name of the network table at the location.}
}
\value{
A data frame giving the networks using the master data table and the nodes they exclude, in
the form of a white space separated string of indices.
}
\description{
Find networks associated with a master data table.
}
\keyword{internal}

