% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/noisyor.R
\name{noisyor}
\alias{noisyor}
\title{Noisy Or constructor}
\usage{
noisyor(index, values, name = as.character(index), parents, pvals,
  params = NULL, params2 = NULL, updateDisambiguation = "T")
}
\arguments{
\item{index}{an integer giving the index of the node}

\item{values}{a vector specifying the values of the node, as strings}

\item{name}{a string specifying the name of the node}

\item{parents}{a vector of integers specifying the indices of the parents of the node}

\item{pvals}{a vector of integers specifying the number of values each parent has}

\item{params}{a matrix specifying the parameters of the node}
}
\value{
a new noisy or node
}
\description{
Constructs a noisy or node
}
\keyword{internal}

