% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{numberOfNetworksUsingDataTable}
\alias{numberOfNetworksUsingDataTable}
\title{Find number of networks associated with a master data table.}
\usage{
numberOfNetworksUsingDataTable(cn, tableinfo)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{dataTable}{The name of the master data table.}

\item{networks}{The name of the network table at the location.}
}
\description{
Find number of networks associated with a master data table.
}
\keyword{internal}

