% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkSerialization.R
\name{parseWSSS_asIntegerVector}
\alias{parseWSSS_asIntegerVector}
\title{Parse W(hite) S(pace) S(eparated) S(trings) as integer vector}
\usage{
parseWSSS_asIntegerVector(str)
}
\arguments{
\item{str}{String to be parsed}
}
\value{
Integer Vector
}
\description{
Parse W(hite) S(pace) S(eparated) S(trings) as integer vector.
}
\keyword{internal}

