% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkSerialization.R
\name{parseWSSS_asStringVector}
\alias{parseWSSS_asStringVector}
\title{Parse W(hite) S(pace) S(eparated) S(trings) as string vector}
\usage{
parseWSSS_asStringVector(str)
}
\arguments{
\item{str}{String to be parsed}
}
\value{
String Vector
}
\description{
Parse W(hite) S(pace) S(eparated) S(trings) as string vector.
}
\keyword{internal}

