% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkSerialization.R
\name{parseWSSS_Dimensions}
\alias{parseWSSS_Dimensions}
\title{Parse W(hite) S(pace) S(eparated) S(trings) for Dimensions}
\usage{
parseWSSS_Dimensions(str)
}
\arguments{
\item{str}{String to be parsed}
}
\value{
6 item integer vector
}
\description{
Parse W(hite) S(pace) S(eparated) S(trings) for Dimensions
}
\keyword{internal}

