% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkSerialization.R
\name{parseWSSS_Params}
\alias{parseWSSS_Params}
\title{Parse W(hite) S(pace) S(eparated) S(trings) Params}
\usage{
parseWSSS_Params(str, dim)
}
\arguments{
\item{str}{String to be parsed.}

\item{dim}{dimension information}
}
\value{
Parameter matrix
}
\description{
Parse W(hite) S(pace) S(eparated) S(trings) for Parameters.
}
\keyword{internal}

