% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DebugWrappers.R
\name{performBatchMIDebug}
\alias{performBatchMIDebug}
\title{A debug wrapper function for performing batch MI.}
\usage{
performBatchMIDebug(net, evidence, nodes, envNodes, weights, number, pause)
}
\description{
A debug wrapper function for performing batch MI. Should be placed in a separate package.
}
\keyword{internal}

