% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ManageExternalInference.R
\name{performMI}
\alias{performMI}
\title{Run the external MI test}
\usage{
performMI(net, evidence, nodes, envNodes, weights, number, filename,
  pause = 2)
}
\arguments{
\item{net}{A network object}

\item{evidence}{The evidence, in 1-indexed, NA=missing form.}

\item{nodes}{The nodes to check mutual information of environment
variables on, in 1-indexed form.}

\item{envNodes}{The environment variables, in 1-indexed form.}

\item{weights}{The weights, if a combined MI result desired. Should be one
weight for each node, or NULL. Intended to be probabilities of nodes,
though importance weights could be used as well.}

\item{number}{The number of samples}

\item{filename}{The full filename to write the results to.}

\item{pause}{The duration in seconds to wait after making system call}

\item{appPath}{The path to the EstimateMI.exe}
}
\value{
Nothing. But writes the MI matrix nodes vs envNodes. If a weights
vector has been given, there will be an additional row which
is a weighted combination of the other rows.
}
\description{
Run the external MI test on a single item passed as evidence
}

