% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DebugWrappers.R
\name{performMIDebug}
\alias{performMIDebug}
\title{A debug wrapper function for performing MI.}
\usage{
performMIDebug(net, evidence, nodes, envNodes, weights, number, pause)
}
\description{
A debug wrapper function for performing MI. Should be placed in a separate package.
}
\keyword{internal}

