% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Importance_Sampling.R
\name{performSampling_importanceSampling}
\alias{performSampling_importanceSampling}
\title{Methods for importance sampling are: nat - sample nodes before
auxiliary samples; or aux - sample auxiliary samples before
nodes.}
\usage{
performSampling_importanceSampling(num, net, evidence, met = "nat")
}
\description{
Methods for importance sampling are: nat - sample nodes before
auxiliary samples; or aux - sample auxiliary samples before
nodes.
}
\keyword{internal}

