% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_bayesnet.R
\name{plot.bayesnet}
\alias{plot.bayesnet}
\title{Plot a network}
\usage{
\method{plot}{bayesnet}(x, ...)
}
\arguments{
\item{x}{the network to plot}

\item{...}{ignored}
}
\description{
Plot a network. If the network has a $layout matrix, this will be used in the
plotting, otherwise the plot will be random. The layout matrix is that used
by igraph and should have
two columns specifying the x and y coordinates of the nodes.
}

