% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/infer_bayesnet.R
\name{predict.bayesnet}
\alias{predict.bayesnet}
\title{Perform inference on network}
\usage{
\method{predict}{bayesnet}(net, evidence, num = 1000, update = FALSE,
  algorithm = "IS", algoSpecific = NULL)
}
\arguments{
\item{net}{the network}

\item{evidence}{vector of integers specifying known values of nodes. This
can include NA values for unknown variables. You can also pass a data frame row
as the function will cast it to numeric internally.}

\item{num}{integer specifying the number of samples to used in inference}

\item{update}{boolean specifying whether to update node parameters}

\item{algorithm}{A string specifying a valid algorithm.
"IS" = Importance sampling.
"BS" = Backwards sampling
"SU" = Selected uniform sampling
"MC" = MCMC sampling}

\item{algoSpecific}{An algorithm specific object:}
}
\value{
updated network with a posterior probabilities object as field
posterior
}
\description{
Perform inference on network
}

