% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prob.R
\name{prob.dirichlet}
\alias{prob.dirichlet}
\title{Calculate probability of value given parents for dirichlet node}
\usage{
\method{prob}{dirichlet}(object, pvals, val, aux, met)
}
\arguments{
\item{pvals}{vector specifying values taken by parents}

\item{val}{integer specifying value taken by node}

\item{dirichlet.object}{the node}
}
\value{
numeric specifying probability of value given parents for node
}
\description{
Calculate probability of value given parents for dirichlet node
}
\keyword{internal}

