% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prob.R
\name{prob.flag}
\alias{prob.flag}
\title{Calculate probability of value given parents for flag node}
\usage{
\method{prob}{flag}(object, pvals, val, aux, met)
}
\arguments{
\item{object}{the node}

\item{pvals}{vector specifying values taken by parents}

\item{val}{integer specifying value taken by node}
}
\value{
numeric specifying probability of value given parents for node
}
\description{
Calculate probability of value given parents for flag node
}
\keyword{internal}

