% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prob.R
\name{prob.flagprod}
\alias{prob.flagprod}
\title{Calculate probability of value given parents for
flag product node}
\usage{
\method{prob}{flagprod}(object, pvals, val, aux, met)
}
\arguments{
\item{object}{the node}

\item{pvals}{vector specifying values taken by parents}

\item{val}{integer specifying value taken by node}

\item{aux}{auxiliary samples}
}
\value{
numeric specifying probability of value given parents for node
}
\description{
Calculate probability of value given parents for
flag product node
}
\keyword{internal}

