% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prob.R
\name{prob.noisyor}
\alias{prob.noisyor}
\title{Get probability noisy or node takes specified value given parents}
\usage{
\method{prob}{noisyor}(object, pvals, val, aux, met)
}
\arguments{
\item{object}{the node object.}

\item{val}{the value taken by the node.}

\item{current.pvals}{the values taken by the node's parents.}
}
\value{
the probability of the node taking the specified value given its
parents take the specified values.
}
\description{
Get probability of a noisy or node take specified value given parents
take specified values. Overrides generic s3 method.
}
\keyword{internal}

