% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prob.R
\name{prob}
\alias{prob}
\title{Get probability of value given parents}
\usage{
prob(object, pvals, val, aux, met = "aux")
}
\arguments{
\item{object}{the node object.}

\item{pvals}{the values taken by the node's parents.}

\item{val}{the value taken by the node.}
}
\value{
the probability of the node taking the specified value given its
parents take the specified values.
}
\description{
S3 generic function for getting the probability of a value given
a nodes parents take specific values.
}
\keyword{internal}

