% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadData.R
\name{readData}
\alias{readData}
\title{Read data (evidence) from a network's data table}
\usage{
readData(cn, net, item, itemTo = NA, zeroindexed = F)
}
\arguments{
\item{net}{The network, with a valid data table name in
the dataTable field - so network tables must have been
initialized.}

\item{item}{The first or only row to read.}

\item{itemTo}{The last row to read, or NA.
at the location.}

\item{zeroindexed}{Specifies if returned values
should be zero indexed.}

\item{connection}{An open RODBC connection}
}
\value{
A dataframe consisting of the specified rows
}
\description{
Read data (evidence) into a network's data table
}

