% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadResults.R
\name{readResults}
\alias{readResults}
\title{Get the posterior distributions for a network for a data item.}
\usage{
readResults(cn, net, item)
}
\arguments{
\item{net}{The network, with a valid results table name in
the resultsTable field - so network tables must have been
initialized.}

\item{item}{The item to retrieve.}

\item{connection}{An open RODBC connection}
}
\value{
A posterior probabilities object
}
\description{
Get the posterior distributions for a network for a data item.
}

