% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{recreateResultTables}
\alias{recreateResultTables}
\title{Recreate result tables for a network}
\usage{
recreateResultTables(cn, net, auto_commit_handled_elsewhere)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{net}{The network.}

\item{auto_commit_handled_elsewhere}{If calling function is handling transaction}
}
\value{
The updated network
}
\description{
Drops the existing result tables and then recreates them
}
\keyword{internal}

