% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RemoveEdge.R
\name{removeEdge}
\alias{removeEdge}
\title{Remove an edge into/a parent to a node}
\usage{
removeEdge(net, child, parent, params = NULL, cn = NULL, deepUpdate = F,
  dropImages = F)
}
\arguments{
\item{child}{The index of the node to remove the parent from.}

\item{parent}{The index of the node to remove as a parent.}

\item{params}{For a multinomial noisy-or node, this is The new parameter 2 matrix.
If NULL a weighted mutex matrix is created. For a dirichlet node, this is the new
parameters matrix.}

\item{cn}{A RODBC connection. Required if performing deep update.}

\item{deepUpdate}{Specifies whether a deep update should be performed.}

\item{dropImages}{Specifies if images should be dropped on a deepUpdate.}

\item{network}{The network}
}
\value{
The updated network
}
\description{
Remove an edge into/a parent to a node. The node must be Noisy-Or or Dirichlet.
}

