% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RemoveEdge.R
\name{removeEdgeDirichlet}
\alias{removeEdgeDirichlet}
\title{Remove an edge into/a parent to a dirichlet node}
\usage{
removeEdgeDirichlet(node, parent, params)
}
\arguments{
\item{node}{The dirichlet node to remove the parent from.}

\item{parent}{The index of the node to remove as a parent.}

\item{params}{The new parameters for the child.}
}
\value{
The updated node
}
\description{
Remove an edge into/a parent to a dirichlet node
}
\keyword{internal}

