% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RemoveEdge.R
\name{removeEdgeNoisyOr}
\alias{removeEdgeNoisyOr}
\title{Remove an edge into/a parent to a noisy-or node}
\usage{
removeEdgeNoisyOr(network, child, parent, params2 = NULL)
}
\arguments{
\item{network}{The network}

\item{child}{The index of the noisy-or node to remove the parent from.}

\item{parent}{The index of the node to remove as a parent.}

\item{params2}{The new parameter 2 matrix for multinomial noisy or nodes.
If NULL a weighted mutex matrix is created.}
}
\value{
The updated network
}
\description{
Remove an edge into/a parent to a noisy-or node
}
\keyword{internal}

