% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RemoveNode.R
\name{removeNode}
\alias{removeNode}
\title{Remove node from network}
\usage{
removeNode(net, index, cn = NULL, deepUpdate = F, dropImages = F)
}
\arguments{
\item{net}{the current network}

\item{index}{The index of the node to remove.}

\item{cn}{A RODBC connection. Required if performing deep update.}

\item{deepUpdate}{Specifies whether a deep update should be performed.}

\item{dropImages}{Specifies if images should be dropped on a deepUpdate.}
}
\value{
Network with node removed.
}
\description{
Remove node from network. Only childless nodes can be
removed. To remove nodes with children, either remove
edges into the node's children (if the children are
Noisy-Or or Dirichlet) or remove the children.
If a layout matrix exists, the row corresponding to the
removed node will also be removed.
NB: A deep update must be performed on a network with a DB backend
}

