% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{restoreAndOverwrite}
\alias{restoreAndOverwrite}
\title{Restore a network and overwrite another}
\usage{
restoreAndOverwrite(cn, overwrite, restore, connection_string = NULL,
  external_path = NULL, dbinfo = dbmsinfo())
}
\arguments{
\item{cn}{A valid ODBC connection.}

\item{overwrite}{The name of the network to be overwritten.}

\item{restore}{The name of the network to be restored.}
}
\value{
A clone of the restored network, with the overwritten network's name.
}
\description{
This function does the following:
(1) The network to be overwritten is loaded and unserialized, without
dropping the master data tables.
(2) The network to be restored is loaded and cloned, with the cloned network having the network to be
overwritten's name.
(3) The cloned network is serialized.
(4) The cloned network is returned.
}

