% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RestoreImage.R
\name{restoreImage}
\alias{restoreImage}
\title{Restore a network's image}
\usage{
restoreImage(cn, net, timeFrom)
}
\arguments{
\item{cn}{An open RODBC connection}

\item{net}{The network. The network must have a valid image table name
in the imageTable field - so network tables must have been initialized.}

\item{timeFrom}{The time stamp of the image to restore}
}
\value{
The restored network
}
\description{
Restore a network's image. NB Due to time stamp resolution in mySQL, storing
multiple images in the same second may cause undefined behavior.
All external inference sessions should be stopped before restoring an image.
}

