% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RestoreImage.R
\name{restoreImage_internal}
\alias{restoreImage_internal}
\title{Restore a network's image}
\usage{
restoreImage_internal(cn, net, timeFrom)
}
\arguments{
\item{cn}{An open RODBC connection}

\item{net}{The network. The network must have a valid image table name
in the imageTable field - so network tables must have been initialized.}

\item{timeFrom}{The time stamp of the image to restore}
}
\value{
The time of the restored image
}
\description{
Restore a network's image. NB Due to time stamp resolution in mySQL, storing
multiple images in the same second may cause undefined behavior. The image is restored
and can be fetched using getNetwork. Note that inf_manager must be copied if retrieved
in this way.
}
\keyword{internal}

