% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TableInfo.R
\name{restoreTableInfo}
\alias{restoreTableInfo}
\title{Unserialize table information object}
\usage{
restoreTableInfo(cn, netRow, dbinfo)
}
\arguments{
\item{cn}{A RODBC connection}

\item{netRow}{Appropriate row from the networks table. The format should be:
Network | DataFamily | DataTables | DataViews | ResultTables |
Images | Datamap | Layout | TableLength}

\item{dbinfo}{A database information object.}
}
\value{
The unserialized new table information object
}
\description{
Unserialize table information object
}
\keyword{internal}

