% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RewindReadapt.R
\name{rewindReadapt}
\alias{rewindReadapt}
\title{Rewind and Readapt given feedback}
\usage{
rewindReadapt(cn, net, from, to, variableIndices, values)
}
\arguments{
\item{cn}{An open RODBC connection}

\item{net}{The network. The network must have a valid image table name
in the imageTable field - so network tables must have been initialized.}

\item{from}{The time from which the data alteration is to occur. This
should be a valid date-time string. For MySQL this is YYYY-MM-DD HH:MM:SS.}

\item{to}{The time to which the data alteration is to occur. This
should be a valid date-time string. For MySQL this is YYYY-MM-DD HH:MM:SS.}

\item{variableIndices}{The indices of the variables
whose values are known in the specified interval.}

\item{values}{The values for the above variables.}
}
\value{
The network set to the state of the appropriate prior image. No readaption will have
yet occurred.
}
\description{
Given user feedback that a particular variable took a particular value
between times t and u this function will:
1. Restore the image closest to but prior to t.
2. Change all data between t and s such that the variable at the specified
   index takes the specified value.
3. Remove all results of items occurring after the time of the
   restored image.
The rewound'  network will be returned. No readaption will have yet occurred.
All external inference sessions should be stopped before running this function.
}

