% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{serializeNetwork}
\alias{serializeNetwork}
\title{Create required entries and tables for a network at a
location using default names.}
\usage{
serializeNetwork(cn, net)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{net}{The network.}
}
\value{
The updated network
}
\description{
Create required entries and tables for a network. Names are as specified in the network's tableinfo
object. The network and it's associated table names will be entered
into the location's network table. The nodes of the network
will be entered into the location's nodes table.
Where the network has the name "net", the the data (input) tables are called "net_data_table_x",
the data (input) views are called "net_data_view_x", the results table is called "net_results",
and the images table is called "net_images".
An initial image will be stored in the image table.
}

