% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RemoveNode.R
\name{shallowUpdateOnNodeRemoved}
\alias{shallowUpdateOnNodeRemoved}
\title{Remove node from network}
\usage{
shallowUpdateOnNodeRemoved(net, index)
}
\arguments{
\item{net}{the current network}

\item{index}{The index of the node to remove.}
}
\value{
Network with node removed.
}
\description{
Remove node from R Bayesnet object
}
\keyword{internal}

