% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RODBCWrapper.R
\name{SqlQuery}
\alias{SqlQuery}
\title{Perform an SQL query}
\usage{
SqlQuery(connection, statement)
}
\arguments{
\item{connection}{An open RODBC connection}

\item{statement}{A valid SQL statement}
}
\value{
A data frame or an invisible -2 if the statement does
not result in returned data.
}
\description{
Perform an SQL query
}
\keyword{internal}

