% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ManageExternalInference.R
\name{startExternalInference}
\alias{startExternalInference}
\title{Start external inference}
\usage{
startExternalInference(net, number = 1000, nextItem = 0, Update = TRUE,
  StoreImages = TRUE, AssumeIntegrity = TRUE,
  Compression_Setting = "None", Compression_Value1 = 10,
  Compression_Value2 = 1, EIE_pause = 500, DBMS_Timeout = 1000,
  pause = 2)
}
\arguments{
\item{net}{The network.}

\item{Update}{Whether network parameters should be updated.}

\item{StoreImages}{Whether images should be stored.}

\item{AssumeIntegrity}{Specifies if engine can assume no other processes will change
database records while running. We recommend you set this to TRUE and act accordingly.}

\item{Compression_Setting}{'None', 'Until' or 'All'. None means compression value 2 is
used through out. All means compression value 1 is used through out. Until will use
value 1 until the process catches up with the number of entries in the data table, and
value 2 after that.}

\item{Compression_Value1}{The number of items to process during an initial catch up}

\item{Compression_Value2}{The number of items to process after any initial catch up}

\item{EIE_pause}{The duration in milliseconds that the external inference engine
will wait before rechecking if new data has appeared in the data table.}

\item{DBMS_Timeout}{Currently unused}

\item{pause}{The duration in seconds to wait after making system call}

\item{Number}{The number of samples per inference case}

\item{NextItem}{The first item to start from. (0-Find Latest)}
}
\value{
Updated network.
}
\description{
Start external inference. Only one external inference engine session can run per
network.
}

