% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ManageExternalInference.R
\name{stopAutomaticInference}
\alias{stopAutomaticInference}
\title{Stop the external inference engine.}
\usage{
stopAutomaticInference(net, pause = 2)
}
\arguments{
\item{pause}{The duration in seconds to wait after making system call}

\item{settings}{A valid externalInferenceSettings object}
}
\value{
Updated network.
}
\description{
Stop the external inference engine.
}

