% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DebugWrappers.R
\name{stopAutomaticInferenceB_debug}
\alias{stopAutomaticInferenceB_debug}
\title{A debug wrapper function for stopping external inference.}
\usage{
stopAutomaticInferenceB_debug(systemID)
}
\description{
A debug wrapper function for stopping external inference. Should be placed in a separate package.
}
\keyword{internal}

