% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StoreImage.R
\name{storeImage}
\alias{storeImage}
\title{Store a network's image}
\usage{
storeImage(cn, net)
}
\arguments{
\item{cn}{An open RODBC connection}

\item{net}{The network. The network must have a valid image table name
in the imageTable field - so network tables must have been initialized.}
}
\description{
Store a network's image. NB Due to time stamp resolution in mySQL, storing
multiple images in the same second may cause undefined behavior.
}

