% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_bayesnet.R
\name{summary.bayesnet}
\alias{summary.bayesnet}
\title{Get summary of network}
\usage{
\method{summary}{bayesnet}(object, ...)
}
\arguments{
\item{object}{The network}

\item{...}{Ignored}
}
\description{
Get summary of network, including:
 class
 size (number of nodes)
 list of names of nodes
 list of posteriori distributions for nodes
Note additional summary arguments ignored.
}

