% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test.R
\name{test.traits}
\alias{test.traits}
\title{Test a constructed node has the expected traits}
\usage{
test.traits(node, params, degree, pindices.jumptable, pvals, parents, index,
  vals)
}
\arguments{
\item{node}{the constructed node.}

\item{params}{a matrix giving the expected parameters for the node.}

\item{degree}{a numeric giving the expected degrees of freedom of the node}

\item{pindices.jumptable}{a vector giving the expected parent indices,
jump table or pmap of the mode, depending on node type.}

\item{pvals}{a vector giving the expected number of values for each parent of the node}

\item{parents}{a vector giving the indices of the parents of the node}

\item{index}{an integer giving the expected index of the node}

\item{vals}{a vector giving the expected values of the node}
}
\value{
vector of boolean values indicating whether above match actual traits
}
\description{
Test a constructed node has the expected traits
}
\keyword{internal}

