% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{unserializeNetwork}
\alias{unserializeNetwork}
\title{Unserialize a network}
\usage{
unserializeNetwork(cn, net, disassociateMasterData, dropMasterData)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{net}{The network.}

\item{disassociateMasterData}{Whether the network should no longer be associated
with its current master data tables.}

\item{dropMasterData}{Specifies whether the master data tables will also be dropped.
If T but other networks still using the master data tables, then they will not be
dropped and a warning will be given. A warning will also be given
if the master data tables are not dropped when no other networks are using them.
This can only be TRUE if disassociateMasterData is also TRUE}
}
\value{
The network, with updated table names.
}
\description{
Remove network entries from networks and nodes tables at a
location, and remove all entries from default tables
for a network at a location. The tables must have been
created by defaultInitializeNetwork.
}

