% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkSerialization.R
\name{unserializeNode}
\alias{unserializeNode}
\title{Unserialize Node}
\usage{
unserializeNode(row)
}
\arguments{
\item{row}{NetworkDataFrame row. See getNetwork}
}
\value{
The unserialized node
}
\description{
Unserialize Node
}
\keyword{internal}

