% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WriteData.R
\name{updateData}
\alias{updateData}
\title{Update data}
\usage{
updateData(cn, net, first, last, variableIndices, values)
}
\arguments{
\item{cn}{An open RODBC connection}

\item{net}{The network}

\item{first}{The first item to delete. All subsequent items will be deleted.}

\item{variableIndices}{The indices of the variables
to change.}

\item{values}{The values to set the specified variables
to.}
}
\description{
Update data in a network's data table by setting a particular variable
to a particular value for a specific set of rows.
}
\keyword{internal}

