% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddNode.R
\name{updateDatamapInDB}
\alias{updateDatamapInDB}
\title{Update datamap in database}
\usage{
updateDatamapInDB(cn, net)
}
\arguments{
\item{cn}{A RODBC connection}

\item{net}{A Bayesian network object}
}
\description{
Update datamap in database
}
\keyword{internal}

