% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddEdge.R
\name{updateImagesOnEdgeAdded}
\alias{updateImagesOnEdgeAdded}
\title{Update images after adding an edge}
\usage{
updateImagesOnEdgeAdded(cn, net, dropImages, from, to)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{net}{The network.}

\item{dropImages}{Whether images should be dropped or updated.}

\item{from}{The index of the parent node}

\item{to}{The index of the child node}
}
\description{
Update the images after adding an edge.
Note that to be able to update the saved network images, the child must be
a binary noisy-or node.
}
\keyword{internal}

