% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RemoveEdge.R
\name{updateImagesOnEdgeRemoved}
\alias{updateImagesOnEdgeRemoved}
\title{Update images after removing an edge}
\usage{
updateImagesOnEdgeRemoved(cn, net, dropImages, to, from)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{net}{The network.}

\item{dropImages}{Whether images should be dropped or updated.}

\item{to}{The index of the child node}

\item{from}{The index of the parent node}
}
\description{
Update the images after removing an edge.
Note that to be able to update the saved network images, the child must be
a binary noisy-or node.
}
\keyword{internal}

