% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AddNode.R
\name{updateImagesOnNodeAdded}
\alias{updateImagesOnNodeAdded}
\title{Update images on node added}
\usage{
updateImagesOnNodeAdded(cn, dropImages, net)
}
\arguments{
\item{cn}{A RODBC connection}

\item{dropImages}{Whether images should simply be dropped}

\item{net}{The network}
}
\description{
Update images on node added.
}
\keyword{internal}

