% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RemoveNode.R
\name{updateImagesOnNodeRemoved}
\alias{updateImagesOnNodeRemoved}
\title{Update images on node removed}
\usage{
updateImagesOnNodeRemoved(cn, dropImages, oldIndex, net)
}
\arguments{
\item{cn}{A RODBC connection}

\item{dropImages}{Whether images should be dropped or altered}

\item{oldIndex}{Index of node that was removed}

\item{net}{The Bayesnet object}
}
\description{
Update images on node removed
}
\keyword{internal}

