% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{updateNetworkLayout}
\alias{updateNetworkLayout}
\title{Update serialized network's layout}
\usage{
updateNetworkLayout(cn, net)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{net}{The network.}

\item{netsTable}{The name of the network table
in the database.}
}
\description{
Updates serialized network's layout.
}

