% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetworkTableManagement.R
\name{updateSerializedNetwork}
\alias{updateSerializedNetwork}
\title{Update serialized network's nodes}
\usage{
updateSerializedNetwork(cn, net, autocommitManagedOutsideScope = F)
}
\arguments{
\item{cn}{An open RODBC connection.}

\item{net}{The network.}

\item{autocommitManagedOutsideScope}{Set to T if you have are controlling the connection's autocommit
outside the scope of this function}

\item{nodesTable}{The name of the nodes table}

\item{networksTable}{The name of the networks table}
}
\description{
Update serialized network's nodes. If nodes have been added or removed you must use the reserialize
function, not the updateSerializedNetwork function.
}
\keyword{internal}

