% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{validate.dirichlet}
\alias{validate.dirichlet}
\title{Validate parameters of dirichlet node}
\usage{
\method{validate}{dirichlet}(object)
}
\description{
Validate parameters of dirichlet node
}
\keyword{internal}

