% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{validate.flagprod}
\alias{validate.flagprod}
\title{Validate parameters of flagprod node}
\usage{
\method{validate}{flagprod}(object)
}
\description{
Validate parameters of flagprod node
}
\keyword{internal}

