% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{validate.logand}
\alias{validate.logand}
\title{Validate inputs for logical and node constructor}
\usage{
\method{validate}{logand}(object)
}
\description{
Validate inputs for logical and node constructor
}
\keyword{internal}

