% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{validate.logprod}
\alias{validate.logprod}
\title{Validate parameters of logprod node}
\usage{
\method{validate}{logprod}(object)
}
\description{
Validate parameters of logprod node
}
\keyword{internal}

