% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{validate.noisyor}
\alias{validate.noisyor}
\title{Validate parameters of noisyor node}
\usage{
\method{validate}{noisyor}(object)
}
\description{
Validate parameters of noisyor node
}
\keyword{internal}

