% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WriteData.R
\name{writeData}
\alias{writeData}
\title{Enter data (evidence) into a network's data table}
\usage{
writeData(cn, evidence, net, zeroindexed = F, time = NULL,
  check_tables = F)
}
\arguments{
\item{cn}{An open RODBC connection}

\item{evidence}{A vector of integers, one for each
variable in the network, specifying the observed value,
or NA for unknown.}

\item{net}{The network, with a valid data table name in
the dataTable field - so network tables must have been
initialized.}

\item{zeroindexed}{Specifies if values are zero indexed.}

\item{time}{Date-time string if you do not want current date time to
be specified. If used it should be in the format YYYY-MM-DD HH:MM:SS.
For combatability with mySQL seconds is the finest grained time permitted.
It is the callers responsibility to ensure that all new entries in the
data table have time stamps later than all previous entries. Failure to
do so will result in undefined behaviour when attempting a rewind-readapt.}

\item{check_tables}{Specifies if the metadata table should be checked to see if
additional data tables have been added to the family. If this is not called, and
such tables have been added, these additional tables will not have evidence
entered into them. Other tables will have, so the correlation of item keys between
tables will have been corrupted. For efficiency, do not set to TRUE if you are
sure no tables could have been added.}
}
\description{
Enter data (evidence) into a network's data table. Note that calling this function when
the network does not include all variables in the data master will lead to such nodes being treated as unknown.
}

